Extract scoring logic out of Autocomplete
